+++ NEWS.md
@@ -11,6 +11,7 @@
 ## Bug fixes
 
 * Fixed issue in `check_collinearity()` for *glmmTMB* models when zero-inflated formula only had one predictor.
+* Fixed issue in `check_model()` for *lme* models.
 
 # performance 0.4.4