+++ NEWS.md
@@ -12,6 +12,7 @@
 
 * Fixed issue in `check_collinearity()` for *glmmTMB* models when zero-inflated formula only had one predictor.
 * Fixed issue in `check_model()` for *lme* models.
+* Fixed issue in `check_heteroscedasticity()` for *aov* objects.
 
 # performance 0.4.4