+++ NEWS.md
@@ -10,6 +10,7 @@
 
 ## Bug fixes
 
+* Fixed issue in `performance_aic()` for models from package *survey*, which returned three different AIC values. Now only the AIC value is returned.
 * Fixed issue in `check_collinearity()` for *glmmTMB* models when zero-inflated formula only had one predictor.
 * Fixed issue in `check_model()` for *lme* models.
 * Fixed issue in `check_heteroscedasticity()` for *aov* objects.