+++ NEWS.md
@@ -14,6 +14,7 @@
 * Fixed issue in `performance_aic()` for models from package *survey*, which returned three different AIC values. Now only the AIC value is returned.
 * Fixed issue in `check_collinearity()` for *glmmTMB* models when zero-inflated formula only had one predictor.
 * Fixed issue in `check_model()` for *lme* models.
+* Fixed issue in `check_distribution()` for *brmsfit* models.
 * Fixed issue in `check_heteroscedasticity()` for *aov* objects.
 * Fixed issues for *lmrob* and *glmrob* objects.